requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
aiohttp>=3.8.0
aiosqlite>=0.17.0